+++ types/rational.go
@@ -253,7 +253,7 @@ func RatEq(t *testing.T, exp, got Rat) (*testing.T, bool, string, Rat, Rat) {
 	return t, exp.Equal(got), "expected:\t%v\ngot:\t\t%v", exp, got
 }
 
-// test if two rat arrays are the equal
+// minimum rational between two
 func MinRat(r1, r2 Rat) Rat {
 	if r1.LT(r2) {
 		return r1